chore: add utils file and add test for locale getter
